requirements.txt
streamlit
pix2tex
pillow
openai
psycopg2-binary